test: re-vendor migrate/integration (stacked) lib into K8s charm
Re-vendor the canonical-postgresql-single-kernel-library at branch
migrate/integration into skl-lib/ and bump the vendored version to
16.2.7 so the lock file tracks the new code. The charm source is
unchanged; only the vendored library and the resolved lock file differ.
