+++ skl-lib/pyproject.toml
@@ -4,7 +4,7 @@
 [project]
 name = "postgresql-charms-single-kernel"
 description = "Shared and reusable code for PostgreSQL-related charms"
-version = "16.2.6"
+version = "16.2.7"
 readme = "README.md"
 license = {file = "LICENSE"}
 authors = [
